requirements.txt
aiogram>=3.7.0
redis>=5.0.0
aiosqlite>=0.19.0
boto3>=1.34.0
python-dotenv>=1.0.0